requirements.txt
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
automium==0.2.6
automium-web==0.1.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.4.1
Flask==3.0.2
h11==0.14.0
idna==3.6
importlib_metadata==7.0.2
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
legacy==0.1.7
MarkupSafe==2.1.5
motor==3.3.2
passlib==1.7.4
Pillow==9.0.0
py-bcrypt==0.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic-validators==0.1.0
pydantic_core==2.16.3
PyJWT==1.7.1
pymongo==4.6.2
pyotp==2.9.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
Werkzeug==3.0.1
zipp==3.17.0
bcrypt==4.1.2
google-api-python-client==2.125.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0